+++ lib/compass/sass_extensions/functions/env.rb
@@ -1,0 +1,5 @@
+module Compass::SassExtensions::Functions::Env
+  def compass_env
+    Sass::Script::String.new((options[:compass][:environment] || "development").to_s)
+  end
+end