+++ test/integrations/compass_test.rb
@@ -73,6 +73,28 @@ class CompassTest < Test::Unit::TestCase
     end
   end
 
+  def test_env_in_development
+    within_project('envtest', lambda {|c| c.environment = :development }) do |proj|
+      each_css_file(proj.css_path) do |css_file|
+        assert_no_errors css_file, 'envtest'
+      end
+      each_sass_file do |sass_file|
+        assert_renders_correctly sass_file, :ignore_charset => true, :environment => "development"
+      end
+    end
+  end
+
+  def test_env_in_production
+    within_project('envtest', lambda {|c| c.environment = :production }) do |proj|
+      each_css_file(proj.css_path) do |css_file|
+        assert_no_errors css_file, 'envtest'
+      end
+      each_sass_file do |sass_file|
+        assert_renders_correctly sass_file, :ignore_charset => true, :environment => "production"
+      end
+    end
+  end
+
   def test_busted_image_urls
     within_project('busted_image_urls') do |proj|
       each_css_file(proj.css_path) do |css_file|
@@ -150,6 +172,7 @@ private
 
     if Compass.configuration.sass_path && File.exists?(Compass.configuration.sass_path)
       compiler = Compass::Compiler.new *args
+      compiler.clean!
       compiler.run
     end
     yield Compass.configuration if block_given?