+++ doc-src/content/CHANGELOG.markdown
@@ -20,6 +20,8 @@ The Documentation for the [latest preview release](http://beta.compass-style.org
 * font-files helper: Stop requiring font type when the type can be guessed from URL
 * inline-font-files: actually works now
 * Upgrade CSS3 Pie to 1.0beta5
+* log sprite generation and removal to the console
+* Added a new helper function `compass-env()` that returns the current compass environment (development, production)
 
 
 0.12.alpha.0 (8/30/2011)